zkvm: add a custom getrandom implementation
This provides a zkvm with the getrandom functionality builtin to the guest
crate. This allows the guest to use the getrandom crate without any
customization.
